Moved scatter utils into utils._code
