+++ mlcolvar/core/loss/committor_loss.py
@@ -17,6 +17,7 @@ __all__ = ["CommittorLoss", "committor_loss", "SmartDerivatives", "compute_descr
 import torch
 import numpy as np
 from typing import Optional
+from mlcolvar.utils._code import scatter_sum
 
 # =============================================================================
 # LOSS FUNCTIONS
@@ -83,38 +84,7 @@ class CommittorLoss(torch.nn.Module):
                                 descriptors_derivatives=self.descriptors_derivatives
                             )
 
-def broadcast(src: torch.Tensor, other: torch.Tensor, dim: int):
-    """Broadcast util, from torch_scatter"""
-    if dim < 0:
-        dim = other.dim() + dim
-    if src.dim() == 1:
-        for _ in range(0, dim):
-            src = src.unsqueeze(0)
-    for _ in range(src.dim(), other.dim()):
-        src = src.unsqueeze(-1)
-    src = src.expand(other.size())
-    return src
 
-def scatter_sum(src: torch.Tensor,
-                index: torch.Tensor,
-                dim: int = -1,
-                out: Optional[torch.Tensor] = None,
-                dim_size: Optional[int] = None) -> torch.Tensor:
-    """Scatter sum function, from torch_scatter module (https://github.com/rusty1s/pytorch_scatter/blob/master/torch_scatter/scatter.py)"""
-    index = broadcast(index, src, dim).to(src.device)
-    if out is None:
-        size = list(src.size())
-        if dim_size is not None:
-            size[dim] = dim_size
-        elif index.numel() == 0:
-            size[dim] = 0
-        else:
-            size[dim] = int(index.max()) + 1
-        out = torch.zeros(size, dtype=src.dtype, device=src.device)
-        return out.scatter_add_(dim, index, src)
-    else:
-        return out.scatter_add_(dim, index, src)
-    
 def committor_loss(x: torch.Tensor, 
                   q: torch.Tensor, 
                   labels: torch.Tensor,
